backend/pyproject.toml
[project]
name = "annotation-agent-backend"
version = "0.1.0"
description = "AI Handwriting Annotation Agent backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pymupdf>=1.24.0",
    "pillow>=10.3.0",
    "opencv-python-headless>=4.9.0.80",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
